Refactor the optimizer
